Rename to data_manager_primer_scheme_bedfiles, add option of input from history
