--- data_managers/data_manager_fetch_sars_cov2_artic_primers/.shed.yml
@@ -1,9 +1,0 @@
-categories:
-- Data Managers
-description: Fetch ARTIC SARS-CoV-2 primers from GitHub and store as BED
-long_description: |
-  This data manager fetches the ARTIC SARS-CoV-2 primer files and installs them in BED format.
-name: data_manager_fetch_artic_sars_cov2_primers
-owner: iuc
-remote_repository_url: https://github.com/galaxyproject/tools-iuc/tree/master/data_managers/data_manager_fetch_artic_sars_cov2_primers
-type: unrestricted
