--- data_managers/data_manager_fetch_sars_cov2_artic_primers/data_manager/fetch_artic_sars_cov2_primers.xml
@@ -1,32 +1,0 @@
-<tool id="data_manager_fetch_artic_sars_cov2_primers" name="ARTIC SARS-CoV-2 primer data manager" version="0.0.7" tool_type="manage_data" profile="19.05">
-    <requirements>
-        <requirement type="package">python</requirement>
-        <requirement type="package" version="2.22.0">requests</requirement>
-    </requirements>
-    <!-- fetch all the primers in one go -->
-    <command detect_errors="exit_code">
-    python '$__tool_directory__/fetch_artic_sars_cov2_primers.py'
-        --galaxy_datamanager_filename '${output_file}'
-        --primers '${primers}'
-    </command>
-    <inputs>
-        <param name="primers" type="select" multiple="true" label="SARS-CoV-2 Primers to fetch">
-            <option value="ARTICv1" selected="true">ARTIC v1</option>
-            <option value="ARTICv2" selected="true">ARTIC v2</option>
-            <option value="ARTICv3" selected="true">ARTIC v3</option>
-        </param>
-    </inputs>
-    <outputs>
-        <data name="output_file" format="data_manager_json"/>
-    </outputs>
-    <tests>
-        <test>
-            <param name="primers" value="ARTICv1,ARTICv2,ARTICv3"/>
-            <output name="output_file">
-                <assert_contents>
-                    <has_text text="ARTIC"/>
-                </assert_contents>
-            </output>        
-        </test>
-    </tests>
-</tool>
