--- data_managers/data_manager_fetch_sars_cov2_artic_primers/tool-data/artic_sars_cov2_primers.loc.sample
@@ -1,7 +1,0 @@
-# this is a tab separated file describing the location of ARTIC primers for use in SARS-CoV-2 sequencing
-#
-# the columns are:
-# value  description path
-#
-# for example
-# ARTICv1   ARTIC v1 primers    /data/galaxy/tool_data/artic_primers/ARTICv1.bed
