--- data_managers/data_manager_fetch_sars_cov2_artic_primers/tool_data_table_conf.xml.sample
@@ -1,7 +1,0 @@
-<tables>
-    <!-- Locations of all ARTIC SARS-CoV-2 primer BED files -->
-    <table name="artic_sars_cov2_primers" comment_char="#">
-        <columns>value, description, path</columns>
-        <file path="tool-data/artic_sars_cov2_primers.loc" />
-    </table>
-</tables>
