+++ data_managers/data_manager_primer_scheme_bedfiles/.shed.yml
@@ -1,0 +1,10 @@
+categories:
+- Data Managers
+description: Install primers schemes from ARTIC SARS-CoV-2  GitHub or history
+long_description: |
+  This data manager install BED format files describing primer schemes, including, but not
+  limited to the ARTIC SARS-CoV-2 primer schemes.
+name: data_manager_primer_scheme_bedfiles
+owner: sanbi-uwc
+remote_repository_url: https://github.com/galaxyproject/tools-iuc/tree/master/data_managers/data_manager_primer_scheme_bedfiles
+type: unrestricted
