+++ data_managers/data_manager_primer_scheme_bedfiles/data_manager/install_primer_scheme_bedfiles.xml
@@ -1,0 +1,64 @@
+<tool id="data_manager_primer_scheme_bedfiles" name="BED-format primer scheme data manager" version="0.0.7" tool_type="manage_data" profile="19.05">
+    <requirements>
+        <requirement type="package">python</requirement>
+        <requirement type="package" version="2.22.0">requests</requirement>
+    </requirements>
+    <!-- fetch all the primers in one go -->
+    <command detect_errors="exit_code">
+    python '$__tool_directory__/install_primer_scheme_bedfiles.py'
+        '${output_file}'
+        #if $input.input_type == "ARTIC"
+            --artic_primers '${input.primers}'
+        #else 
+            --primer_file '${input.primer_input}'
+            --primer_name '${input.primer_name}'
+            --primer_description '${input.primer_description}'
+        #end if
+    </command>
+    <inputs>
+        <conditional name="input">
+            <param name="input_type" label="Choose the source for primer schemes" type="select">
+                <option value="ARTIC" selected="true">ARTIC SARS-CoV-2 Github page</option>
+                <option value="history">History</option>
+            </param>
+            <when value="ARTIC">                    
+                <param name="primers" type="select" multiple="true" label="SARS-CoV-2 Primers to fetch">
+                    <option value="ARTICv1" selected="true">ARTIC v1</option>
+                    <option value="ARTICv2" selected="true">ARTIC v2</option>
+                    <option value="ARTICv3" selected="true">ARTIC v3</option>
+                </param>
+            </when>
+            <when value="history">
+                <param name="primer_input" label="Select history item" type="data" format="bed" />
+                <param name="primer_name" label="Name for this primer scheme" type="text" 
+                       help="Non-word characters will be removed from primer scheme name and spaces replaced with underscore (_)" />
+                <param name="primer_description" label="Description for primer scheme" type="text" />
+            </when>
+        </conditional>
+    </inputs>
+    <outputs>
+        <data name="output_file" format="data_manager_json"/>
+    </outputs>
+    <tests>
+        <test>
+            <param name="input_type" value="ARTIC" />
+            <param name="primers" value="ARTICv1,ARTICv2,ARTICv3"/>
+            <output name="output_file">
+                <assert_contents>
+                    <has_text text="ARTIC"/>
+                </assert_contents>
+            </output>        
+        </test>
+        <test>
+            <param name="input_type" value="history" />
+            <param name="primer_input" ftype="bed" value="sample1.bed" />
+            <param name="primer_name" value="sample_primer" />
+            <param name="primer_description" value="sample primer scheme" />
+            <output name="output_file">
+                <assert_contents>
+                    <has_text text="sample_primer"/>
+                </assert_contents>
+            </output>        
+        </test>
+    </tests>
+</tool>
