+++ data_managers/data_manager_primer_scheme_bedfiles/data_manager/test-data/sample1.bed
@@ -1,0 +1,7 @@
+MN908947.3      30      54      nCoV-2019_1_LEFT        60      +
+MN908947.3      385     410     nCoV-2019_1_RIGHT       60      -
+MN908947.3      320     342     nCoV-2019_2_LEFT        60      +
+MN908947.3      704     726     nCoV-2019_2_RIGHT       60      -
+MN908947.3      642     664     nCoV-2019_3_LEFT        60      +
+MN908947.3      1004    1028    nCoV-2019_3_RIGHT       60      -
+MN908947.3      943     965     nCoV-2019_4_LEFT        60      +
