+++ data_managers/data_manager_primer_scheme_bedfiles/data_manager_conf.xml
@@ -1,16 +1,17 @@
 <?xml version="1.0"?>
 <data_managers>    
-    <data_manager tool_file="data_manager/fetch_artic_sars_cov2_primers.xml" id="fetch_artic_sars_cov2_primers">
-        <data_table name="artic_sars_cov2_primers">
+    <data_manager tool_file="data_manager/install_primer_scheme_bedfiles.xml" id="data_manager_primer_scheme_bedfiles">
+        <data_table name="primer_scheme_bedfiles">
             <output>
                 <column name="value" />
                 <column name="description" />
                 <column name="path" output_ref="output_file" >
+                    <!-- note: the Python script sanitises the possibly user-supplied scheme name ('value') -->
                     <move type="file">
                         <source>${path}</source>
-                        <target base="${GALAXY_DATA_MANAGER_DATA_PATH}">artic_sars_cov2_primers/#echo str($value)#.bed</target>
+                        <target base="${GALAXY_DATA_MANAGER_DATA_PATH}">primer_scheme_bedfiles/#echo str($value)#.bed</target>
                     </move>
-                    <value_translation>artic_sars_cov2_primers/#echo str($value)#.bed</value_translation>
+                    <value_translation>primer_scheme_bedfiles/#echo str($value)#.bed</value_translation>
                     <value_translation type="function">abspath</value_translation>
                 </column>
             </output>
