+++ data_managers/data_manager_primer_scheme_bedfiles/tool-data/primer_scheme_bedfiles.loc.sample
@@ -1,0 +1,8 @@
+# this is a tab separated file describing the location of primer schemes used for amplicon
+# sequencing (e.g. using the ARTIC SARS-CoV-2 protocol)
+#
+# the columns are:
+# value  description path
+#
+# for example
+# ARTICv1   ARTIC v1 primers    /data/galaxy/tool_data/artic_primers/ARTICv1.bed
