+++ data_managers/data_manager_primer_scheme_bedfiles/tool_data_table_conf.xml.sample
@@ -1,0 +1,7 @@
+<tables>
+    <!-- Locations of primer scheme BED files -->
+    <table name="primer_scheme_bedfiles" comment_char="#">
+        <columns>value, description, path</columns>
+        <file path="tool-data/primer_scheme_bedfiles.loc" />
+    </table>
+</tables>
